moving stylesheets, adding pointer event
