+++ src/stylesheets/LandingPage.scss
@@ -79,6 +79,7 @@
 
     .eventDetails {
         margin-top: 5%;
+        justify-content: space-evenly;
     }
 }
